[B/ref] Modify post creation API to retrieve email from token

+++ src/main/java/com/cantata/tradetalent/domain/Post/dto/request/UploadPostRequest.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 @Builder
 public class UploadPostRequest {
 
-    private Long id;  // 자동 생성된 id를 저장할 필드
+    private int id;  // 자동 생성된 id를 저장할 필드
 
     @JsonProperty("email")
     String email;
